requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
cachetools==6.2.4
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==46.0.3
css==0.1
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.8
django-channels==0.7.0
django-cors-headers==4.3.1
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-genai
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hypothesis==6.148.7
idna==3.10
jiter==0.11.1
numpy==2.4.0
oauthlib==3.3.1
openai==2.6.0
packaging==25.0
pillow==11.1.0
platformdirs==4.3.8
proto-plus==1.27.0
protobuf==5.29.5
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
virtualenv==20.31.2
whitenoise==6.9.0
pinecone-client==5.0.1
tenacity==8.2.3
python-docx
torch
sentence-transformers
